services/API_gateway_service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.23.1
pydantic==2.4.0
requests==2.31.0
psycopg2-binary==2.9.9